+++ apps/sim/app/workspace/[workspaceId]/logs/components/index.ts
@@ -1,6 +1,5 @@
 export { Dashboard } from './dashboard'
 export { LogDetails, LogDetailsContent } from './log-details'
-export { ExecutionSnapshot } from './log-details/components/execution-snapshot'
 export { FileCards } from './log-details/components/file-download'
 export { TraceView } from './log-details/components/trace-view'
 export { LogRowContextMenu } from './log-row-context-menu'
