--- apps/sim/app/workspace/[workspaceId]/logs/components/log-details/components/execution-snapshot/index.ts
@@ -1,1 +1,0 @@
-export { ExecutionSnapshot } from './execution-snapshot'
