+++ apps/sim/lib/api/contracts/index.ts
@@ -1,3 +1,11 @@
+/**
+ * Deliberately NOT re-exported from this barrel: `./tools` (per-integration tool
+ * contracts), `./selectors`, `./v1` (admin API), and `./demo-requests`. All of their
+ * consumers import those files directly, and re-exporting them here shipped their Zod
+ * schema construction (~60 KB gzip) to every route that touches any contract — schema
+ * objects are built at module scope, so `export *` defeats tree-shaking for them.
+ * Import from the specific contract file instead.
+ */
 export * from './admin'
 export * from './api-keys'
 export * from './audit-logs'
@@ -7,7 +15,6 @@ export * from './cli-auth'
 export * from './common'
 export * from './copilot'
 export * from './credentials'
-export * from './demo-requests'
 export * from './desktop-auth'
 export * from './desktop-tool-authorization'
 export * from './environment'
@@ -23,15 +30,12 @@ export * from './primitives'
 export * from './sandboxes'
 export * from './secret-mount-policy'
 export * from './secrets'
-export * from './selectors'
 export * from './skills'
 export * from './storage-transfer'
 export * from './subscription'
 export * from './tool-primitives'
-export * from './tools'
 export * from './types'
 export * from './user'
-export * from './v1'
 export * from './workflows'
 export * from './workspace-file-folders'
 export * from './workspace-files'
